+++ src/graphic_context.hpp
@@ -87,10 +87,10 @@ public:
 
   virtual void draw_string_centered(float x, float y, const std::string& str, Color color = Color (0xFFFFFFFF)) =0;
 
-  virtual void set_clip_rect (int x1_, int y1_, int x2_, int y2_) =0;
+  virtual void set_clip_rect (float x1_, float y1_, float x2_, float y2_) =0;
 
-  virtual int get_width ()  =0;
-  virtual int get_height () =0;
+  virtual float get_width ()  =0;
+  virtual float get_height () =0;
 
   virtual void clear () =0;
