+++ src/gui_file_button.hpp
@@ -34,8 +34,8 @@ public:
   void on_mouse_enter() override { m_mouse_over = true; }
   void on_mouse_leave() override { m_mouse_over = false; }
 
-  void on_primary_button_press(int x, int y) override;
-  void on_primary_button_release(int x, int y) override;
+  void on_primary_button_press(float x, float y) override;
+  void on_primary_button_release(float x, float y) override;
 
 protected:
   std::string m_filename;
